requirements.txt
torch>=2.2
faiss-cpu
einops
numpy
scikit-learn
tqdm